requirements.txt
aiohttp
asyncio
dropbox
chardet
python-dotenv
pytest